Move Go tests out of `pkg/codegen`
This commit moves the Go tests into
`pkg/codegen/testing/test/go.go`, so that they won't be included in
`pkg/codegen` code by downstream dependencies.
