+++ pkg/codegen/go/gen_program_test/batch2/gen_program_test.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2022, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -18,13 +18,20 @@ package batch2
 
 import (
 	"os"
+	"path/filepath"
 	"testing"
+
+	"github.com/stretchr/testify/require"
 
 	codegen "github.com/pulumi/pulumi/pkg/v3/codegen/go"
 	"github.com/pulumi/pulumi/pkg/v3/codegen/testing/test"
 )
 
 func TestGenerateProgram(t *testing.T) {
-	os.Chdir("../../../go") // chdir into codegen/go
-	codegen.GenerateProgramBatchTest(t, test.ProgramTestBatch(2, 6))
+	rootDir, err := filepath.Abs(filepath.Join("..", "..", "..", "..", ".."))
+	require.NoError(t, err)
+
+	// Change into pkg/codegen/go
+	os.Chdir(filepath.Join(rootDir, "pkg", "codegen", "go"))
+	test.GenerateProgramBatchTest("go")(t, rootDir, codegen.GenerateProgram, test.ProgramTestBatch(2, 6))
 }
