+++ pkg/codegen/go/gen_program_test/batchyaml/gen_program_test.go
@@ -15,14 +15,12 @@
 package batchyaml
 
 import (
-	"os"
 	"path/filepath"
 	"testing"
 
 	"github.com/hashicorp/hcl/v2"
 	"github.com/stretchr/testify/assert"
 
-	"github.com/pulumi/pulumi/pkg/v3/codegen"
 	codegenGo "github.com/pulumi/pulumi/pkg/v3/codegen/go"
 	"github.com/pulumi/pulumi/pkg/v3/codegen/pcl"
 	"github.com/pulumi/pulumi/pkg/v3/codegen/testing/test"
@@ -33,25 +31,16 @@ import (
 func TestGenerateProgram(t *testing.T) {
 	t.Parallel()
 
-	sdkDir, err := filepath.Abs(filepath.Join("..", "..", "..", "..", "..", "sdk"))
+	rootDir, err := filepath.Abs(filepath.Join("..", "..", "..", "..", ".."))
 	assert.NoError(t, err)
 
-	err = os.Chdir("../../../go") // chdir into codegen/go
-	assert.NoError(t, err)
-
-	test.TestProgramCodegen(t,
-		test.ProgramCodegenOptions{
-			Language:   "go",
-			Extension:  "go",
-			OutputFile: "main.go",
-			Check: func(t *testing.T, path string, dependencies codegen.StringSet) {
-				codegenGo.Check(t, path, dependencies, sdkDir)
-			},
-			GenProgram: func(program *pcl.Program) (map[string][]byte, hcl.Diagnostics, error) {
-				// Prevent tests from interfering with each other
-				return codegenGo.GenerateProgramWithOptions(program,
-					codegenGo.GenerateProgramOptions{ExternalCache: codegenGo.NewCache()})
-			},
-			TestCases: test.PulumiPulumiYAMLProgramTests,
-		})
+	test.GenerateGoYAMLBatchTest(
+		t,
+		rootDir,
+		func(program *pcl.Program) (map[string][]byte, hcl.Diagnostics, error) {
+			// Prevent tests from interfering with each other
+			return codegenGo.GenerateProgramWithOptions(program,
+				codegenGo.GenerateProgramOptions{ExternalCache: codegenGo.NewCache()})
+		},
+	)
 }
